refactor: rename store folder to (store) to more clearly identify page segments
