+++ src/app/[locale]/request-offers/(components)/abandon-request-dialog.tsx
@@ -18,7 +18,7 @@ import {
 import { Button } from "@/components/ui/button";
 import { useRouter } from "@/i18n/routing";
 
-import { useClearRequestOffersStore } from "../store/request-offers-store-provider";
+import { useClearRequestOffersStore } from "../(store)/request-offers-store-provider";
 
 export function AbandonRequestDialog({ ...props }: ComponentProps<typeof AlertDialogTrigger>) {
    const t = useTranslations();
