+++ src/app/[locale]/request-offers/(components)/use-step-manager.tsx
@@ -4,7 +4,7 @@ import { useEffect } from "react";
 
 import { Pathname, usePathname, useRouter } from "@/i18n/routing";
 
-import { useLastUnlockedStep, useSetStepVerified } from "../store/request-offers-store-provider";
+import { useLastUnlockedStep, useSetStepVerified } from "../(store)/request-offers-store-provider";
 
 const basePath = "/request-offers";
 export const stepToPath: Record<number, Pathname> = {
